# src/main/resources/META-INF/resources/api.yml
openapi: 3.0.1
info:
  title: Transaction Processing Service
  version: 1.0.0
  description: Transaction Processing Service
  x-custom-info:
    authors:
      name: Dasun Anushka
      email: lakmal.thanthrige@101digital.io
servers:
  - url: 'http://localhost:8090'
paths:
  /configurations:
    post:
      operationId: addConfigurations
      summary: Create configurations
      tags:
        - Configurations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Configurations'
      responses:
        '200':
          description: Confirm success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigurationsResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected error while performing file operations.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      operationId: getConfigurations
      summary: Get configurations
      tags:
        - Configurations
      parameters:
        - name: type
          required: false
          in: query
          schema:
            type: string
          example: ROUNDUP_CONTRIBUTION
        - name: walletId
          required: true
          in: query
          schema:
            type: string
          example: 18957-94517-15684-46645
        - name: pageNumber
          required: false
          in: query
          schema:
            type: int
          example: 1
        - name: pageSize
          required: false
          in: query
          schema:
            type: int
          example: 10
        - name: listOrders
          in: query
          description: Results will be sorted and ordered based on the given criteria
          required: false
          schema:
            type: string
          example: createdAt-DESC,updatedAt-ASC
      responses:
        '200':
          description: Confirm success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigurationsListResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected error while performing file operations.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearer: [ ]


  /configurations/{configurationId}:
    delete:
      tags:
        - Configurations
      summary: Delete the configuration using configuration id
      operationId: deleteConfigurationById
      parameters:
        - name: configurationId
          in: path
          description: configuration id
          required: true
          schema:
            type: string
          example: 123
      responses:
        503:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        204:
          description: Contact deleted successfully
          content:
            application/json:
              schema:
                $ref: string
        404:
          description: Data Not Found
          content:
            application/json:
              schema:
                type: string
        400:
          description: Input data is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - Bearer: [ ]

#-----
components:
  schemas:
    TestResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Test'
        paging:
          $ref: '#/components/schemas/PagingInformation'
    Test:
      type: object
      properties:
        name:
          type: string
    PagingInformation:
      type: object
      properties:
        totalRecords:
          type: integer
          description: Number of total records in the response.
          format: int64
          example: 100
        pageNumber:
          type: integer
          description: The page number to return (starting at page 1).
          format: int32
          example: 1
        pageSize:
          type: integer
          description: The number of records to return in a single request
          format: int32
          example: 10
    ErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ResponseStatus'
    ResponseStatus:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
    Configurations:
      type: object
      properties:
        id:
          type: UUID
        type:
          type: string
        walletId:
          type: string
        logicCode:
          type: string
        createdBy:
          type: UUID
        updatedby:
          type: UUID
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        supplementaryData:
          $ref: '#/components/schemas/OBSupplementaryData1'
    OBSupplementaryData1:
      type: object
      description: Additional information that can not be captured in the structured fields and/or any other specific block.
    ConfigurationsResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Configurations'
    ConfigurationsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Configurations'
        paging:
          $ref: '#/components/schemas/PagingInformation'
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer